+++ ampel/ztf/t0/ArchiveUpdater.py
@@ -107,11 +107,13 @@ class ArchiveUpdater(ArchiveDBClient):
 
         # add healpix index
         candidate = alert["candidate"]
-        candidate["_hpx"] = lonlat_to_healpix(
-            candidate["ra"] * u.deg,
-            candidate["dec"] * u.deg,
-            self.NSIDE,
-            order="nested",
+        candidate["_hpx"] = int(
+            lonlat_to_healpix(
+                candidate["ra"] * u.deg,
+                candidate["dec"] * u.deg,
+                self.NSIDE,
+                order="nested",
+            )
         )
 
         insert_alert = postgresql.insert(Alert).values(
